extract DemoArea
