+++ templates/demo/sandbox/main.scss
@@ -41,8 +41,5 @@ body {
   width: 100%;
   padding: 1em;
   flex-basis: auto;
-
-  // pre {
-  //   width: 50%;
-  // }
+  box-sizing: border-box;
 }
